add crete to disoute

+++ OctagonPlatform/Models/Transaction.cs
@@ -9,24 +9,25 @@ namespace OctagonPlatform.Models
     [NotMapped]
     public class Transaction
     {
-            public string SequenceNumber { get; set; }
-            public int Amount1 { get; set; }
-            public float Amount2 { get; set; }
-            public object Miscelanea1 { get; set; }
-            public object Miscelanea2 { get; set; }
-            public string Miscelaneax { get; set; }
-            public string Pan { get; set; }
-            public string DateExp { get; set; }
-            public string ServiceCode { get; set; }
-            public int Id { get; set; }
-            public string CommIdentifier { get; set; }
-            public string Terminalidentifier { get; set; }
-            public string SoftwareVersion { get; set; }
-            public string ModeFlag { get; set; }
-            public string Infoheader { get; set; }
-            public string TerminalId { get; set; }
-            public string TransCode { get; set; }
-            public DateTime Date { get; set; }
-            public string MessageType { get; set; }
+        public string SequenceNumber { get; set; }
+        public int Amount1 { get; set; }
+        public double Amount2 { get; set; }
+        public object Miscelanea1 { get; set; }
+        public object Miscelanea2 { get; set; }
+        public string Miscelaneax { get; set; }
+        public string Pan { get; set; }
+        public string DateExp { get; set; }
+        public string ServiceCode { get; set; }
+        public int Id { get; set; }
+        public string CommIdentifier { get; set; }
+        public string Terminalidentifier { get; set; }
+        public string SoftwareVersion { get; set; }
+        public string ModeFlag { get; set; }
+        public string Infoheader { get; set; }
+        public string TerminalId { get; set; }
+        public string TransCode { get; set; }
+        public DateTime Date { get; set; }
+        public string MessageType { get; set; }
     }
+
 }
